Move the dispatch logoutUser test from Nav into App to match the new placement of the action
